Removed format strings
